requirements.txt
requests~=2.32.3
pandas~=2.2.2
haversine~=2.8.1
joblib~=1.4.2
matplotlib~=3.9.1
numpy~=1.26.4
optuna~=3.6.1
tensorflow~=2.17.0
filterpy~=1.4.5
prophet~=1.1.5
scikit-learn~=1.5.1
statsmodels~=0.14.2
xgboost~=2.1.0
fastapi~=0.111.1
uvicorn~=0.30.3
psycopg2-binary~=2.9.9
sqlalchemy~=2.0.31
alembic~=1.13.2
python-dotenv~=1.0.1
pydantic~=2.8.2
keras~=3.4.1
cachetools~=5.4.0
pyjwt~=2.8.0
passlib~=1.7.4
python-jose~=3.3.0
cryptography~=43.0.0
asyncpg~=0.29.0
plotly~=5.23.0
jose~=1.0.0
beautifulsoup4~=4.12.3
pillow~=10.4.0
Faker~=26.1.0
asyncio~=3.4.3
pycryptodome~=3.20.0
firebase-admin~=6.5.0
bcrypt~=4.2.0
redis~=5.1.0b7
networkx~=3.3
fastapi-cache2~=0.2.2
geopy~=2.4.1
folium~=0.17.0
googlemaps~=4.10.0
osmnx~=1.9.4
polyline~=2.0.2